Fix the css post processing by gulp.
The uncss plugin relies on the final html files being available in order
to detect what css to include or not. But the sass task, which included
the uncss-ing (in production builds), runs before the pelican task has
begun, so there are no html files at that point.

Added a separate function to be run after the pelican build which post
processes the css and renames the final css file.